# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Member {
  id: number;
  name: string;
  initials: string;
  color: string;
  role: string;
  online: boolean;
}

export type Priority = (typeof Priority)[keyof typeof Priority];

export const Priority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type Status = (typeof Status)[keyof typeof Status];

export const Status = {
  todo: "todo",
  inprogress: "inprogress",
  review: "review",
  done: "done",
} as const;

export interface Task {
  id: number;
  title: string;
  description: string;
  priority: Priority;
  status: Status;
  /** @nullable */
  assigneeId: number | null;
  /** @nullable */
  dueDate: string | null;
  tags: string[];
  position: number;
  commentCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface Comment {
  id: number;
  taskId: number;
  /** @nullable */
  authorId: number | null;
  authorName: string;
  text: string;
  createdAt: string;
}

export type TaskDetail = Task & {
  comments: Comment[];
};

export interface CreateTaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: Priority;
  status: Status;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  tags?: string[];
}

export interface UpdateTaskInput {
  /** @minLength 1 */
  title?: string;
  description?: string;
  priority?: Priority;
  status?: Status;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  tags?: string[];
}

export interface MoveTaskInput {
  status: Status;
  position?: number;
}

export interface CreateCommentInput {
  /** @minLength 1 */
  text: string;
  /** @nullable */
  authorId?: number | null;
}

export interface StatusBreakdown {
  todo: number;
  inprogress: number;
  review: number;
  done: number;
}

export interface PriorityBreakdown {
  high: number;
  medium: number;
  low: number;
}

export interface BoardSummary {
  total: number;
  completionRate: number;
  overdue: number;
  dueSoon: number;
  completedThisWeek: number;
  byStatus: StatusBreakdown;
  byPriority: PriorityBreakdown;
}

export type ActivityKind = (typeof ActivityKind)[keyof typeof ActivityKind];

export const ActivityKind = {
  created: "created",
  moved: "moved",
  commented: "commented",
  completed: "completed",
} as const;

export interface Activity {
  id: string;
  kind: ActivityKind;
  taskId: number;
  taskTitle: string;
  actorName: string;
  actorColor: string;
  message: string;
  createdAt: string;
}

export interface MemberWorkload {
  member: Member;
  total: number;
  active: number;
  done: number;
  overdue: number;
}

export type ListTasksParams = {
  search?: string;
  status?: ListTasksStatus;
  assigneeId?: number;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  todo: "todo",
  inprogress: "inprogress",
  review: "review",
  done: "done",
} as const;

export type GetBoardActivityParams = {
  limit?: number;
};