# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: members
    description: Team members
  - name: tasks
    description: Tasks and Kanban operations
  - name: comments
    description: Task comments
  - name: board
    description: Board summary and activity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List team members
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
            enum: [todo, inprogress, review, done]
          required: false
        - in: query
          name: assigneeId
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task with its comments
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/move:
    patch:
      operationId: moveTask
      tags: [tasks]
      summary: Move a task to a different status column
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveTaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  /board/summary:
    get:
      operationId: getBoardSummary
      tags: [board]
      summary: Get high-level board statistics
      responses:
        "200":
          description: Board summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardSummary"

  /board/activity:
    get:
      operationId: getBoardActivity
      tags: [board]
      summary: Recent activity feed across the board
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          required: false
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  /board/workload:
    get:
      operationId: getBoardWorkload
      tags: [board]
      summary: Workload distribution by team member
      responses:
        "200":
          description: Workload by member
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberWorkload"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        initials:
          type: string
        color:
          type: string
        role:
          type: string
        online:
          type: boolean
      required: [id, name, initials, color, role, online]

    Priority:
      type: string
      enum: [high, medium, low]

    Status:
      type: string
      enum: [todo, inprogress, review, done]

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/Status"
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        tags:
          type: array
          items:
            type: string
        position:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - priority
        - status
        - assigneeId
        - dueDate
        - tags
        - position
        - commentCount
        - createdAt
        - updatedAt

    Comment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        authorId:
          type: ["integer", "null"]
        authorName:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, taskId, authorId, authorName, text, createdAt]

    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
          required: [comments]

    CreateTaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          default: ""
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/Status"
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        tags:
          type: array
          items:
            type: string
          default: []
      required: [title, priority, status]

    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/Status"
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        tags:
          type: array
          items:
            type: string

    MoveTaskInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/Status"
        position:
          type: integer
      required: [status]

    CreateCommentInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
        authorId:
          type: ["integer", "null"]
      required: [text]

    StatusBreakdown:
      type: object
      properties:
        todo:
          type: integer
        inprogress:
          type: integer
        review:
          type: integer
        done:
          type: integer
      required: [todo, inprogress, review, done]

    PriorityBreakdown:
      type: object
      properties:
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
      required: [high, medium, low]

    BoardSummary:
      type: object
      properties:
        total:
          type: integer
        completionRate:
          type: number
        overdue:
          type: integer
        dueSoon:
          type: integer
        completedThisWeek:
          type: integer
        byStatus:
          $ref: "#/components/schemas/StatusBreakdown"
        byPriority:
          $ref: "#/components/schemas/PriorityBreakdown"
      required:
        - total
        - completionRate
        - overdue
        - dueSoon
        - completedThisWeek
        - byStatus
        - byPriority

    Activity:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [created, moved, commented, completed]
        taskId:
          type: integer
        taskTitle:
          type: string
        actorName:
          type: string
        actorColor:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - kind
        - taskId
        - taskTitle
        - actorName
        - actorColor
        - message
        - createdAt

    MemberWorkload:
      type: object
      properties:
        member:
          $ref: "#/components/schemas/Member"
        total:
          type: integer
        active:
          type: integer
        done:
          type: integer
        overdue:
          type: integer
      required: [member, total, active, done, overdue]