# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityKind } from "./activityKind";

export interface Activity {
  id: string;
  kind: ActivityKind;
  taskId: number;
  taskTitle: string;
  actorName: string;
  actorColor: string;
  message: string;
  createdAt: Date;
}