# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  taskId: number;
  /** @nullable */
  authorId: number | null;
  authorName: string;
  text: string;
  createdAt: Date;
}