# lib/api-zod/src/generated/types/moveTaskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { Status } from "./status";

export interface MoveTaskInput {
  status: Status;
  position?: number;
}