# lib/api-zod/src/generated/types/priorityBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */

export interface PriorityBreakdown {
  high: number;
  medium: number;
  low: number;
}