# lib/api-zod/src/generated/types/status.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */

export type Status = (typeof Status)[keyof typeof Status];

export const Status = {
  todo: "todo",
  inprogress: "inprogress",
  review: "review",
  done: "done",
} as const;