# lib/api-zod/src/generated/types/statusBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */

export interface StatusBreakdown {
  todo: number;
  inprogress: number;
  review: number;
  done: number;
}