# lib/api-zod/src/generated/types/updateTaskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { Priority } from "./priority";
import type { Status } from "./status";

export interface UpdateTaskInput {
  /** @minLength 1 */
  title?: string;
  description?: string;
  priority?: Priority;
  status?: Status;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: Date | null;
  tags?: string[];
}