# lib/db/src/schema/members.ts
import { pgTable, serial, text, boolean } from "drizzle-orm/pg-core";

export const membersTable = pgTable("members", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  initials: text("initials").notNull(),
  color: text("color").notNull(),
  role: text("role").notNull().default("Contributor"),
  online: boolean("online").notNull().default(true),
});

export type Member = typeof membersTable.$inferSelect;
export type InsertMember = typeof membersTable.$inferInsert;